move creating and maintenance of formation assignments to formation assaignment service
